Update document for auth
